+++ spec/rspec/mocks/any_instance_spec.rb
@@ -884,6 +884,166 @@ module RSpec
         end
       end
 
+      context "passing the receiver to the implementation block" do
+        context "when configured to pass the instance" do
+          include_context 'with isolated configuration'
+          before(:each) do
+            RSpec::Mocks.configuration.yield_receiver_to_any_instance_implementation_blocks = true
+          end
+
+          describe "an any instance stub" do
+            it "passes the instance as the first arg of the implementation block" do
+              instance = klass.new
+
+              expect { |b|
+                klass.any_instance.should_receive(:bees).with(:sup, &b)
+                instance.bees(:sup)
+              }.to yield_with_args(instance, :sup)
+            end
+
+            it "does not pass the instance to and_call_original" do
+              klass = Class.new do
+                def call(*args)
+                  args.first
+                end
+              end
+              klass.any_instance.should_receive(:call).and_call_original
+              instance = klass.new
+              expect(instance.call(:bees)).to be :bees
+            end
+          end
+
+          describe "an any instance expectation" do
+            it "doesn't effect with" do
+              instance = klass.new
+              klass.any_instance.should_receive(:bees).with(:sup)
+              instance.bees(:sup)
+            end
+
+            it "passes the instance as the first arg of the implementation block" do
+              instance = klass.new
+
+              expect { |b|
+                klass.any_instance.should_receive(:bees).with(:sup, &b)
+                instance.bees(:sup)
+              }.to yield_with_args(instance, :sup)
+            end
+          end
+        end
+
+        context "when configured not to pass the instance" do
+          include_context 'with isolated configuration'
+          before(:each) do
+            RSpec::Mocks.configuration.yield_receiver_to_any_instance_implementation_blocks = false
+          end
+
+          describe "an any instance stub" do
+            it "does not pass the instance to the implementation block" do
+              instance = klass.new
+
+              expect { |b|
+                klass.any_instance.should_receive(:bees).with(:sup, &b)
+                instance.bees(:sup)
+              }.to yield_with_args(:sup)
+            end
+
+            it "does not cause with to fail when the instance is passed" do
+              instance = klass.new
+              klass.any_instance.should_receive(:bees).with(:faces)
+              instance.bees(:faces)
+            end
+          end
+        end
+
+        context "by default" do
+          let(:block_regex) { /as the first block argument/ }
+
+          it "is off" do
+            expect(RSpec::Mocks.configuration.yield_receiver_to_any_instance_implementation_blocks?).to be false
+          end
+
+          it "will warn about allowances receiving blocks in 3.0" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            allow_any_instance_of(klass).to receive(:foo) { |args| }
+            klass.new(:faces).foo
+          end
+
+          it "will warn about expectations receiving blocks in 3.0" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            expect_any_instance_of(klass).to receive(:foo) { |args| }
+            klass.new(:faces).foo
+          end
+
+          it "will warn about expectations receiving blocks with a times restriction" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            klass.any_instance.should_receive(:foo).exactly(3).times { :some_return_value }
+
+            instance = klass.new(:faces)
+            3.times { instance.foo }
+          end
+
+          it "will warn about expectations receiving blocks with an argument expectation" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            klass.any_instance.should_receive(:foo).with(3) { :some_return_value }
+
+            instance = klass.new(:faces)
+            instance.foo(3)
+          end
+
+          it "works with a do end style block" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            klass.any_instance.should_receive(:foo).with(3) do
+              :some_return_value
+            end
+
+            instance = klass.new(:faces)
+            instance.foo(3)
+          end
+
+          it "won't warn if there is no implementation block on an expectation" do
+            expect(RSpec).not_to receive(:warn_deprecation)
+
+            klass = Struct.new(:bees)
+
+            allow_any_instance_of(klass).to receive(:foo)
+            klass.new(:faces).foo
+          end
+
+          it "will warn about stubs receiving blocks in 3.0" do
+            expect(RSpec).to receive(:warn_deprecation).with(block_regex)
+
+            klass = Struct.new(:bees)
+
+            expect_any_instance_of(klass).to receive(:foo) { |args| }
+            klass.new(:faces).foo
+          end
+
+          it "won't warn if there is no implementation block on an stub" do
+            expect(RSpec).not_to receive(:warn_deprecation)
+
+            klass = Struct.new(:bees)
+
+            allow_any_instance_of(klass).to receive(:foo)
+            klass.new(:faces).foo
+          end
+        end
+      end
+
       context 'when used in conjunction with a `dup`' do
         it "doesn't cause an infinite loop" do
           pending "This intermittently fails on JRuby" if RUBY_PLATFORM == 'java'